Refactor data handling and improve performance

+++ include/dsa/MergeSort.hpp
@@ -10,22 +10,22 @@ void mergeSort(Array<T> &arr, Func func, bool reverse = false)
     if (arr.size() <= 1)
         return;
 
-    int mid = arr.size() / 2;
+    size_t mid = arr.size() / 2;
 
     Array<T> left;
-    for(int i = 0; i < mid; i++){
+    for(size_t i = 0; i < mid; i++){
         left.append(arr[i]);
     }
 
     Array<T> right;
-    for(int i = mid; i < arr.size(); i++){
+    for(size_t i = mid; i < arr.size(); i++){
         right.append(arr[i]);
     }
 
     mergeSort(left, func, reverse);
     mergeSort(right, func, reverse);
 
-    int i = 0, j = 0, k = 0;
+    size_t i = 0, j = 0, k = 0;
 
     while (i < left.size() && j < right.size())
     {
